task4/pdf-summarizer/pyproject.toml
[project]
name = "pdf-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.6.1",
    "pypdf2>=3.0.1",
    "streamlit>=1.51.0",
]
